user/model update
