+++ src/server/users/models.js
@@ -1,0 +1,41 @@
+import db from '../../server/database';
+import {KeyUtils, ENTITY} from '../../server/key-utils';
+
+export class CreatedPostManager {
+  static addCreatedPost(entity, entityKey, userKey, timeHash, cb) {
+    console.log('model');
+    if (entity !== ENTITY.ITEM && entity !== ENTITY.IMAGE) {
+      return cb(new Error(('invalid entity'), null));
+    }
+    const idxKey = KeyUtils.getIdxKey(ENTITY.CREATED_POST, timeHash, userKey);
+    const idxPost = {
+      entity,
+      key: entityKey
+    };
+    console.log(idxKey);
+    console.log(idxPost);
+    return db.put(idxKey, idxPost, (err) => {
+      if (err) {
+        return cb(new Error('error while putting in DB'), null);
+      }
+      return cb(null, idxKey);
+    });
+  }
+}
+export class SavedPostManager {
+  static addSavedPost(entityKey, userKey, timeHash, cb) {
+    const idxKey = KeyUtils.getIdxKey(ENTITY.SAVED_POST, timeHash, userKey);
+    const idxPost = {
+      key: entityKey
+    };
+    return db.put(idxKey, idxPost, (err) => {
+      if (err) {
+        return cb(new Error('error while putting in DB'), null);
+      }
+      return cb(null, idxKey);
+    });
+  }
+}
+
+
+
