+++ src/test/node/index.js
@@ -11,8 +11,7 @@ import './items/models';
 import './images/models';
 import './images/controllers';
 import './users/controllers';
-import './users/createdPostsModel';
-import './users/savedPostsModel';
+import './users/models';
 import './database';
 import './aws-s3';
 import './auth-token';
